Koordinat tablosu güncelleme

+++ database/migrations/2019_05_15_092104_create_user_coordinates_table.php
@@ -16,8 +16,8 @@ class CreateUserCoordinatesTable extends Migration
         Schema::create('user_coordinates', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('user_id');
-            $table->float('latitude');
-            $table->float('longitude');
+            $table->double('latitude');
+            $table->double('longitude');
             $table->timestamps();
         });
     }
